+++ CMakeLists.txt
@@ -95,6 +95,7 @@ add_library(libninja OBJECT
 	src/metrics.cc
 	src/parser.cc
 	src/state.cc
+	src/status.cc
 	src/string_piece_util.cc
 	src/util.cc
 	src/version.cc
